# backend/docs/swagger.yaml
basePath: /
definitions:
  handlers.AccountResponse:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  handlers.ApiErrorResponse:
    properties:
      error:
        type: string
    required:
    - error
    type: object
...
  handlers.DeleteAccountRequest:
    properties:
      password:
        type: string
    required:
    - password
    type: object
  handlers.MessageResponse:
    properties:
      message:
        type: string
    required:
    - message
    type: object
...
  handlers.UpdateEmailRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handlers.UpdatePasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        type: string
    required:
    - current_password
    - new_password
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: A music management API with artists and albums
  title: Better Music API
  version: "1.0"
paths:
  /api/account:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Current password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.DeleteAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
      security:
      - Bearer: []
      summary: Permanently delete the current account
      tags:
      - account
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AccountResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
      security:
      - Bearer: []
      summary: Get the current account
      tags:
      - account
  /api/account/email:
    put:
      consumes:
      - application/json
      parameters:
      - description: New email and current password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
      security:
      - Bearer: []
      summary: Change the current account email
      tags:
      - account
  /api/account/password:
    put:
      consumes:
      - application/json
      parameters:
      - description: Current and new passwords
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdatePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
      security:
      - Bearer: []
      summary: Change the current account password
      tags:
      - account
...
schemes:
- http
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"